Merging The 3 APIs
* DataProcess API

The API between Data Storage System and the Compute Engine. It focuses on reading and checking the input given by the Compute Engine. Then the written output will be sent back and will be different depending if the program has recently computed the same input.

* User to Compute Engine API

An API that allows the data to be sent to and from the code and the user via a network

* Simple API for ComputeEngine

It is still a work in progress but the essentials of this API is communicating data that will be sent or received from other parts of the entire software system