Checkpoint3TestSuite Fixing (#18)
* Update CP1Tester.java

* fail2

* working 2.1

* fail 1

* Update CP1Tester.java

* Update CP1Tester.java

* Update CP1Tester.java

* DataProcess API

The API between Data Storage System and the Compute Engine. It focuses on reading and checking the input given by the Compute Engine. Then the written output will be sent back and will be different depending if the program has recently computed the same input.

* Pushing API again

* Fix Attempt1

* Make it more simple

* fix annotations

* Update DataCheck.java

* Delete CP1Tester.java

* making filler constructors

* i think i found it

* trying to fix it

* Update ProcessAPIProto.java

* more stuff

* something

* Update build.gradle

* Switch Package name

* last fix

* local config file for gradle properties

* Update build.gradle

* Update TestStatusCheckPR.java

* Original DataProcess API I created

After fixing the gradle and checkstyle I want my original API to work as well

* DataProcess API

* processAPI fix

* User to Compute Engine API

An API that allows the data to be sent to and from the code and the user via a network

* Simple API for ComputeEngine

It is still a work in progress but the essentials of this API is communicating data that will be sent or received from other parts of the entire software system

* Simple Implementations

(Made this a while ago but forgot to commit)

* Changing back API names

Github did not like it when I changed the names of the files

* Update ProcessAPIProto.java

* Update DataProcess.java

* Update DataProcess.java

* Update DataProcess.java

* Update DataProcess.java

* Update DataProcess.java

* Create Input and Output General Configuration

* SmokeTest and InMemoryConfig

Fixed my APIs and the methods I need from then. Also completed the Implementations and SmokeTest as well as added the InputConfig and OutputConfig. The InMemory is also completed.

* ChecklistFix1

* ChecktyleFix2

* Checkstyle fix3

* Checkstyle fix 4

* Checkstyle Fix 4

* Checkstyle fix  5

* FilterTests 1

* FilterTests 2

* FilterTest 3

* FilterTest 4

* Filter5 + FixTests

* Last Filter Test

* CP3 complete1

* fix name error

* Dumb fix

* fix4

* a

* Suite Fix

* Attmempt to fix testsuite3

* FixingTestSuite3

* restarting Suit tests